Revert "Merge branch 'feature/GEODE-7665' into feature/GEODE-7678"
This reverts commit 96dc078f90fa6cd62586ad655391df1114407bb0, reversing
changes made to 8db8c47383774e841e4ad737cbec567a6c10d5e5.

+++ geode-core/src/main/java/org/apache/geode/internal/cache/partitioned/rebalance/model/PartitionedRegionLoadModel.java
@@ -362,9 +362,6 @@ public class PartitionedRegionLoadModel {
    * to create a bucket on that node. Because the bucket operator is asynchronous, the bucket may
    * not be created immediately, but the model will be updated regardless. Invoke
    * {@link #waitForOperations()} to wait for those operations to actually complete
-   *
-   * @param bucket the bucket for which a redundant copy should be made
-   * @param targetMember the member on which a redundant copy of a bucket should be made
    */
   public void createRedundantBucket(final BucketRollup bucket, final Member targetMember) {
     Map<String, Long> colocatedRegionSizes = getColocatedRegionSizes(bucket);
@@ -756,8 +753,8 @@ public class PartitionedRegionLoadModel {
   /**
    * For testing only, calculate the total variance of the members
    */
-  public double getVarianceForTest() {
-    double variance = 0;
+  public float getVarianceForTest() {
+    float variance = 0;
 
     for (Member member : this.members.values()) {
       variance += variance(member.getTotalLoad(), member.getWeight(), getAverageLoad());
@@ -769,8 +766,8 @@ public class PartitionedRegionLoadModel {
   /**
    * For testing only, calculate the total variance of the members
    */
-  public double getPrimaryVarianceForTest() {
-    double variance = 0;
+  public float getPrimaryVarianceForTest() {
+    float variance = 0;
 
     for (Member member : this.members.values()) {
       variance += variance(member.getPrimaryLoad(), member.getWeight(), getPrimaryAverage());
